api_yamdb/requirements.txt
<<<<<<< HEAD
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2022.7.1
sqlparse==0.4.3

asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.0.12
colorama==0.4.6
Django==3.2
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
idna==3.4
iniconfig==2.0.0
mypy==1.0.1
mypy-extensions==1.0.0
packaging==23.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.10.0
pydocstyle==6.3.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.7.1
requests==2.26.0
snowballstemmer==2.2.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
python-dotenv==0.21.1
=======
asgiref==3.6.0
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1 
python-dotenv==0.21.1
pytest-django==4.4.0
>>>>>>> 7d9f817691c6611eb87e60d5e77a69843c7166ab
